Add consistency check for syslinux_patch() callers
If SYSLINUX was unable to patch both ldlinux.sys and the boot sector by
calling syslinux_patch() function, it will only print the error messages
out on the screen and continue the installation process instead of
aborting the installation process.

Now we check the return value of each syslinux_patch() call to make sure
whether we can go on the installation process or not.

Signed-off-by: Paulo Alcantara <pcacjr@gmail.com>